+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/auth-stack-builder/auth-stack-transform.ts
@@ -160,7 +160,6 @@ export class AmplifyAuthTransform extends AmplifyCategoryTransform {
     let cognitoStackProps: CognitoStackOptions = {
       ...this._cliInputs.cognitoConfig,
       ...roles,
-      // eslint-disable-next-line spellcheck/spell-checker
       breakCircularDependency: FeatureFlags.getBoolean('auth.breakcirculardependency'),
       // eslint-disable-next-line spellcheck/spell-checker
       useEnabledMfas: FeatureFlags.getBoolean('auth.useenabledmfas'),