+++ packages/amplify-cli/src/plugin-manager.ts
@@ -219,7 +219,7 @@ export const addPluginPackage = async (pluginPlatform: PluginPlatform, pluginDir
 
 // remove: select from the plugins only,
 // if the location belongs to the scan directories, put the info inside the excluded.
-// if the location is in the useraddedlocaitons, remove it from the user added locations.
+// if the location is in the userAddedLocations, remove it from the user added locations.
 export function removePluginPackage(pluginPlatform: PluginPlatform, pluginInfo: PluginInfo): void {
   // remove from the plugins
   if (pluginPlatform.plugins[pluginInfo.manifest.name] && pluginPlatform.plugins[pluginInfo.manifest.name].length > 0) {