+++ packages/amplify-graphql-migration-tests/CHANGELOG.md
@@ -37,6 +37,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [2.4.0-beta.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.4.0-beta.0...amplify-graphql-migration-tests@2.4.0-beta.1) (2022-11-19)
+## [2.3.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.3.5...amplify-graphql-migration-tests@2.3.6) (2023-01-17)
 
 **Note:** Version bump only for package amplify-graphql-migration-tests
 
@@ -56,6 +57,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [2.4.0-cdkv2.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.4.0-cdkv2.0...amplify-graphql-migration-tests@2.4.0-cdkv2.3) (2022-11-10)
+## [2.3.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.3.4...amplify-graphql-migration-tests@2.3.5) (2023-01-03)
 
 **Note:** Version bump only for package amplify-graphql-migration-tests