+++ packages/amplify-util-headless-input/CHANGELOG.md
@@ -28,6 +28,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 ## [1.9.7-beta.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-headless-input@1.9.6...amplify-util-headless-input@1.9.7-beta.0) (2022-11-17)
+## [1.9.8](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-headless-input@1.9.7...amplify-util-headless-input@1.9.8) (2023-01-03)
 
 **Note:** Version bump only for package amplify-util-headless-input