+++ packages/amplify-util-import/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 # [2.3.0-beta.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@2.3.0-beta.0...amplify-util-import@2.3.0-beta.3) (2022-12-28)
+## [2.3.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@2.3.0...amplify-util-import@2.3.1) (2023-01-03)
 
 **Note:** Version bump only for package amplify-util-import
 
@@ -52,6 +53,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [2.3.0-cdkv2.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@2.2.36...amplify-util-import@2.3.0-cdkv2.0) (2022-11-01)
+# [2.3.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@2.2.36...amplify-util-import@2.3.0) (2022-10-27)
 
 
 ### Features